requirements.txt
asgiref==3.6.0
Django==4.1.5
django-filter==22.1
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.2.1
mongoengine==0.25.0
pymongo==3.12.3
python-snappy==0.6.1
pytz==2022.7
sqlparse==0.2.4
tzdata==2022.7
winkerberos==0.9.1
